+++ client/src/App.js
@@ -13,15 +13,13 @@ import Cart from "./pages/Cart/Cart";
 import PageNotFound from "./pages/PageNotFound/PageNotFound";
 import {useDispatch, useSelector} from "react-redux";
 import ProductModal from "./components/ProductModal/ProductModal";
-import {fetchAsyncProducts} from './store/productsSlice'
+import {fetchAsyncProducts, fetchAsyncBestSellers, fetchAsyncTrending} from './store/productsSlice'
 import {
     fetchAsyncChairs,
     fetchAsyncLamps,
     fetchAsyncDecor,
     fetchAsyncFurniture,
     fetchAsyncSofas,
-    fetchAsyncTrending,
-    fetchAsyncBestSellers
 } from './store/productsFiltersSlice'
 import Error from './components/Error/Error'
 import Loader from './components/Loader/Loader'
@@ -37,6 +35,15 @@ function App() {
     useEffect(() => {
         dispatch(fetchAsyncProducts())
     }, [dispatch])
+
+    useEffect(() => {
+        dispatch(fetchAsyncTrending())
+    }, [dispatch])
+
+    useEffect(() => {
+        dispatch(fetchAsyncBestSellers())
+    }, [dispatch])
+
 
     useEffect(() => {
         dispatch(fetchAsyncChairs())
@@ -58,13 +65,6 @@ function App() {
         dispatch(fetchAsyncSofas())
     }, [dispatch])
 
-    useEffect(() => {
-        dispatch(fetchAsyncTrending())
-    }, [dispatch])
-
-    useEffect(() => {
-        dispatch(fetchAsyncBestSellers())
-    }, [dispatch])
 
 
     switch (status) {
