+++ client/src/components/Bestsellers/Bestsellers.jsx
@@ -6,22 +6,11 @@ import {getElement, openModal} from "../../store/productsSlice";
 
 function Bestsellers(props) {
     const dispatch = useDispatch()
-  const products = useSelector((state) => state.products.products);
-  const filteredItems = products.filter((el) => el.bestSeller);
-
-// const [filteredItems, setFilteredItems] = useState({})
-//     useEffect(() => {
-//         fetch('http://localhost:3001/api/products/filter?trendingProduct=true')
-//             .then(res => res.json())
-//             .then(list => {
-//                 setFilteredItems(list)
-//             })
-//     }, [])
-//     console.log(filteredItems)
     function handleProductClick(product) {
         dispatch(getElement(product));
         dispatch(openModal())
     }
+    const bestSellers = useSelector(state => state.products.bestSellers)
 
 
     return (
@@ -29,7 +18,7 @@ function Bestsellers(props) {
       <h2>Best Sellers Products</h2>
       <p>Top sale in this week</p>
       <div className={styles.flexWrapper}>
-        {filteredItems.map((item) => (
+        {bestSellers.products.map((item) => (
           <BestsellerItem key={item._id} item={item}  onClick={()=>handleProductClick(item._id)} />
         ))}
       </div>
